+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - `tinymist` is now bundled in plugin.
 - Add TextMate grammar.
+- Add support of semantic highlighting
 
 ## [0.0.7]